Moved CRUD for users to Client/Worker controllers and updated frontend code to route correctly.

+++ api/Services/ClientService.cs
@@ -1,4 +1,5 @@
 
+using System.Diagnostics.Eventing.Reader;
 using api.DAL;
 using api.DTO;
 using api.Models;
@@ -14,11 +15,36 @@ public class ClientService : IClientService
         _repository = repository;
         _logger = logger;
     }
-
-    public async Task<IEnumerable<ClientDto>> GetAll() // Get all clients
+    private bool IsAuthorized(Client client, string? authUserId, string? role)
     {
+        if (string.IsNullOrEmpty(authUserId))
+        {
+            _logger.LogWarning("[ClientService] Authorization failed: AuthUserId is null or empty.");
+            return false;
+        } 
+        var ok = false;
+
+        if (role == "Admin") return true;
+
+        else if (role == "Client" && client.AuthUserId == authUserId)
+        {
+            ok = true;
+        }
+        return ok;
+    }
+    public async Task<IEnumerable<ClientDto>> GetAll(bool isAdmin) // Get all clients
+    {
+        if (!isAdmin) // If not admin, return empty
+        {
+            _logger.LogWarning("[ClientService] Unauthorized access attempt to get all clients by non-admin user.");
+            return Enumerable.Empty<ClientDto>();
+        }
         var clients = await _repository.GetAll();
-        if (clients == null || !clients.Any()) return Enumerable.Empty<ClientDto>();
+        if (clients == null || !clients.Any())
+        {
+            _logger.LogWarning("[ClientService] No clients found.");
+            return Enumerable.Empty<ClientDto>();
+        } 
 
         var clientDtos = clients.Select(c => new ClientDto
         {
@@ -33,13 +59,18 @@ public class ClientService : IClientService
 
     }
 
-    public async Task<ClientDto?> GetByAuthUserId(string authUserId) // Get client by AuthUserId
-    {
+    public async Task<ClientDto?> GetByAuthUserId(string authUserId, string authId, string role) // Get client by AuthUserId
+    {   
         var client = await _repository.GetByAuthUserId(authUserId); // Get client from repository
         if (client == null)
         {
             _logger.LogWarning("[ClientService] Client not found for AuthUserId {AuthUserId}", authUserId);
             return null;
+        }
+        if(!IsAuthorized(client, authId, role))
+        {
+            _logger.LogWarning("[ClientService] Unauthorized access attempt by AuthUserId {AuthUserId} to delete ClientId {ClientId:0000}", authId, client.Id);
+            throw new UnauthorizedAccessException("You are not authorized to delete this client.");
         }
 
         var clientDto = new ClientDto // Map Client to ClientDto
@@ -51,10 +82,11 @@ public class ClientService : IClientService
             Email = client.Email,
             AuthUserId = client.AuthUserId
         };
-        return clientDto; // return null if not found
+        return clientDto; // return empty if not found
     }
 
-    public async Task<ClientDto?> GetById(int id) // Get client by Id
+    //add authentication
+    public async Task<ClientDto?> GetById(int id, string authUserId, string role) // Get client by Id
     {
         var client = await _repository.GetClientById(id); // Get client from repository
         if (client == null)
@@ -62,6 +94,11 @@ public class ClientService : IClientService
             _logger.LogWarning("[ClientService] Client not found for Id {Id:0000}", id);
             return null;
         }
+        if(!IsAuthorized(client, authUserId, role))
+        {
+            _logger.LogWarning("[ClientService] Unauthorized access attempt by AuthUserId {AuthUserId} to get ClientId {ClientId:0000}", authUserId, id);
+            throw new UnauthorizedAccessException("You are not authorized to get this client.");
+        }
 
         var clientDto = new ClientDto // Map Client to ClientDto
         {
@@ -72,10 +109,10 @@ public class ClientService : IClientService
             Email = client.Email,
             AuthUserId = client.AuthUserId
         };
-        return clientDto; // return null if not found
+        return clientDto; // return empty if not found
     }
-    
-    public async Task<ClientDto> Create(ClientDto dto) // Create new client
+
+    public async Task<ClientDto> Create(RegisterDto dto, string authId) // Create new client
     {
         var client = new Client // Map ClientDto to Client
         {
@@ -83,13 +120,13 @@ public class ClientService : IClientService
             Address = dto.Address,
             Phone = dto.Phone,
             Email = dto.Email,
-            AuthUserId = dto.AuthUserId
+            AuthUserId = authId
         };
 
         bool created = await _repository.Create(client); // Create client in repository
         if (!created)
         {
-            _logger.LogError("[ClientService] Client creation failed {@client}", client);
+            _logger.LogWarning("[ClientService] Client creation failed {@client}", client);
             throw new InvalidOperationException("Create failed");
         }
 
@@ -106,13 +143,20 @@ public class ClientService : IClientService
 
         return createdDto; // return created client dto
     }
-    public async Task<bool> Update(UpdateUserDto userDto) // Update existing client
+
+    //Add authorization
+    public async Task<bool> Update(UpdateUserDto userDto, string authId, string role) // Update existing client
     {
         var id = userDto.Id; // Get client Id from dto
         var existingClient = await _repository.GetClientById(id); // Get existing client from repository
         if (existingClient == null)
         {
             return false;
+        }
+        if(!IsAuthorized(existingClient, authId, role))
+        {
+            _logger.LogWarning("[ClientService] Unauthorized access attempt by AuthUserId {AuthUserId} to delete ClientId {ClientId:0000}", authId, id);
+            throw new UnauthorizedAccessException("You are not authorized to delete this client.");
         }
 
         existingClient.Name = userDto.Name; // Update client properties
@@ -123,27 +167,29 @@ public class ClientService : IClientService
         bool updated = await _repository.Update(existingClient); // Update client in repository
         if (!updated)
         {
-            _logger.LogError("[ClientService] Update failed for HealthcareWorkerId {HealthcareWorkerId:0000}, {@worker}", id, existingClient);
+            _logger.LogWarning("[ClientService] Update failed for HealthcareWorkerId {HealthcareWorkerId:0000}, {@worker}", id, existingClient);
             throw new InvalidOperationException($"Update operation failed for HealthcareWorkerId {id}");
         }
         return updated; // return true if updated
     }
-    
-    public async Task<bool> Delete(int id) // Delete client by Id
+
+    //add authorization
+    public async Task<bool> Delete(int id, string authId, string role) // Delete client by Id
     {
         var client = await _repository.GetClientById(id); // Get client from repository
         if (client is null) return false; // normal "not found"
-
+        if(!IsAuthorized(client, authId, role))
+        {
+            _logger.LogWarning("[ClientService] Unauthorized access attempt by AuthUserId {AuthUserId} to delete ClientId {ClientId:0000}", authId, id);
+            throw new UnauthorizedAccessException("You are not authorized to delete this client.");
+        }
         var ok = await _repository.Delete(id); // Delete client in repository
         if (!ok)
         {
-            _logger.LogError("[ClientService] Client deletion failed for Id {Id:0000}", id);
+            _logger.LogWarning("[ClientService] Client deletion failed for Id {Id:0000}", id);
             throw new InvalidOperationException($"Delete operation failed for client ID {id}");
 
         }
-
         return true;
     }
-
-
 }
